+++ uliweb/contrib/secretkey/__init__.py
@@ -2,7 +2,7 @@ from uliweb import settings, functions
 from uliweb.utils.common import import_attr, application_path
 from hashlib import md5
 
-def get_cipher(key=None):
+def get_cipher(key=None, keyfile=None):
     """
     Get cipher object, and then you can invoke:
         des = get_cipher()
@@ -13,36 +13,37 @@ def get_cipher(key=None):
     kwargs = settings.SECRETKEY.CIPHER_ARGS
     
     if not key:
-        key = functions.get_cipher_key()
+        key = functions.get_cipher_key(keyfile)
     cipher = des_func(key, **kwargs)
     return cipher
 
-def encrypt(v, key=None):
+def encrypt(v, key=None, keyfile=None):
     """
     Encrypt an string
     """
-    cipher = functions.get_cipher(key)
+    cipher = functions.get_cipher(key, keyfile)
     return cipher.encrypt(v)
 
-def decrypt(v, key=None):
+def decrypt(v, key=None, keyfile=None):
     """
     Encrypt an string
     """
-    cipher = functions.get_cipher(key)
+    cipher = functions.get_cipher(key, keyfile)
     return cipher.decrypt(v)
 
-def get_key():
+def get_key(keyfile=None):
     """
     Read the key content from secret_file
     """
-    with file(application_path(settings.SECRETKEY.SECRET_FILE), 'rb') as f:
+    keyfile = keyfile or application_path(settings.SECRETKEY.SECRET_FILE)
+    with file(keyfile, 'rb') as f:
         return f.read()
     
-def get_cipher_key():
+def get_cipher_key(keyfile=None):
     """
     Create key which will be used in des, because des need 8bytes chars
     """
-    _key = get_key()
+    _key = get_key(keyfile)
     _k = md5(_key).hexdigest()
     key = xor(_k[:8], _k[8:16], _k[16:24], _k[24:])
     return key